requirements.txt
flask==2.0.1
gunicorn==20.1.0
redis==4.0.2
rq==1.10.1
tensorflow-cpu==2.7.0
numpy==1.21.4
pandas==1.3.4
python-dotenv==0.19.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.36
alembic==1.7.7
pgvector==0.2.1
